fix: reject initialize, ping, and notifications/initialized methods in the new protocol session
